backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
fastapi
uvicorn
python-dotenv
google-generativeai==0.5.4

